+++ app/gui/trading/__init__.py
@@ -1,5 +1,6 @@
+from .commodities_tab import CommoditiesTab
 from .trading_tab import TradingTab
 from .uex_trading_tab import UEXTradingTab
 
 
-__all__ = ["TradingTab", "UEXTradingTab"]
+__all__ = ["CommoditiesTab", "TradingTab", "UEXTradingTab"]